libsql: move push_with_retry to sync.rs
